+++ test/transform/resource/after-delombok/NonNullOnParameterOfDefaultMethod.java
@@ -3,7 +3,7 @@ interface NonNullOnParameterOfDefaultMethod {
 	void test(@lombok.NonNull String arg);
 	default void test2(@lombok.NonNull String arg) {
 		if (arg == null) {
-			throw new java.lang.NullPointerException("arg is marked @NonNull but is null");
+			throw new java.lang.NullPointerException("arg is marked non-null but is null");
 		}
 		System.out.println(arg);
 	}