+++ test/transform/resource/after-delombok/NonNullTypeUse.java
@@ -2,12 +2,12 @@ import lombok.NonNull;
 class NonNullTypeUse {
 	void test1(@NonNull String[][][] args) {
 		if (args == null) {
-			throw new java.lang.NullPointerException("args is marked @NonNull but is null");
+			throw new java.lang.NullPointerException("args is marked non-null but is null");
 		}
 	}
 	void test2(String @NonNull [][][] args) {
 		if (args == null) {
-			throw new java.lang.NullPointerException("args is marked @NonNull but is null");
+			throw new java.lang.NullPointerException("args is marked non-null but is null");
 		}
 	}
 	void test3(String[] @NonNull [][] args) {
@@ -16,17 +16,17 @@ class NonNullTypeUse {
 	}
 	void test5(@NonNull String simple) {
 		if (simple == null) {
-			throw new java.lang.NullPointerException("simple is marked @NonNull but is null");
+			throw new java.lang.NullPointerException("simple is marked non-null but is null");
 		}
 	}
 	void test6(java.lang.@NonNull String weird) {
 		if (weird == null) {
-			throw new java.lang.NullPointerException("weird is marked @NonNull but is null");
+			throw new java.lang.NullPointerException("weird is marked non-null but is null");
 		}
 	}
 	void test7(java.lang.String @NonNull [][] weird) {
 		if (weird == null) {
-			throw new java.lang.NullPointerException("weird is marked @NonNull but is null");
+			throw new java.lang.NullPointerException("weird is marked non-null but is null");
 		}
 	}
 }