+++ test/transform/resource/after-delombok/NonNullWithSneakyThrows.java
@@ -2,7 +2,7 @@ class NonNullWithSneakyThrows {
 	void test(@lombok.NonNull String in) {
 		try {
 			if (in == null) {
-				throw new java.lang.NullPointerException("in is marked @NonNull but is null");
+				throw new java.lang.NullPointerException("in is marked non-null but is null");
 			}
 			System.out.println(in);
 		} catch (final java.lang.Throwable $ex) {