+++ test/transform/resource/after-delombok/SetterOnClass.java
@@ -54,7 +54,7 @@ class SetterOnClass6 {
 	@java.lang.SuppressWarnings("all")
 	public void setNonNull(@lombok.NonNull final String nonNull) {
 		if (nonNull == null) {
-			throw new java.lang.NullPointerException("nonNull is marked @NonNull but is null");
+			throw new java.lang.NullPointerException("nonNull is marked non-null but is null");
 		}
 		this.nonNull = nonNull;
 	}